Create README.md
Photo Gallery App built with React and Firebase. It’s similar to Instagram in its layout and functionality. React was used for the front end. And for the back end Firebase storage is used to store images and Firebase firestore, a database is used to keep track of image urls and give real time image data making refreshing the page unnecessary after adding a new image.